refactor: use `find_program` for `chameleon` and rename some things
