+++ doc/troubleshooting.md
@@ -28,3 +28,13 @@ Then rebuild (`meson compile` and/or `meson install`).
 
 Remember to revert this change and rebuild/re-install, so that Iguana runs with
 full optimization when you are processing large data sets (`-Dbuildtype=release`).
+
+### 🔵 I got some error about "chameleon", or an error in some "chameleon" file that I can't find
+
+[Chameleon is a code generator](/src/chameleon) to automatically create
+`iguana` bindings for programming languages other than C++. All generated code
+is produced in your build directory. If you have issues with Chameleon, either:
+- an `Action.yaml` file is not correct
+- something is wrong with `chameleon`
+
+In either case, open an issue or contact the maintainers.
