+++ src/chameleon/README.md
@@ -1,6 +1,12 @@
 # Chameleon
 
-Given a `yaml` action function specification, generate:
+Given a `yaml` action function specification, `chameleon` generates:
 - language bindings
 - action function tests, for all languages
-- `Algorithm::Run` functions
+
+`chameleon` is run automatically by `meson` at build-time; all generated
+code is produced in the build tree. If you want to test `chameleon`, you
+may also run it locally:
+```bash
+chameleon --help
+```
